requirements.txt
Pillow
numpy
tensorflow
mediapipe
